Split QueryContext into base and http variants
