+++ tests/inc/editor/configuration-loader/ConfigurationLoaderTest.php
@@ -3,7 +3,7 @@
 namespace RemoteDataBlocks\Editor;
 
 use PHPUnit\Framework\TestCase;
-use RemoteDataBlocks\Config\QueryContext;
+use RemoteDataBlocks\Config\HttpQueryContext;
 use RemoteDataBlocks\Test\TestDatasource;
 
 class ConfigurationLoaderTest extends TestCase {
@@ -12,7 +12,7 @@ class ConfigurationLoaderTest extends TestCase {
 	}
 
 	public function testIsRegisteredBlockReturnsTrueForRegisteredBlock() {
-		ConfigurationLoader::register_block( 'some_slick_block', new QueryContext( new TestDatasource() ) );
+		ConfigurationLoader::register_block( 'some_slick_block', new HttpQueryContext( new TestDatasource() ) );
 		$this->assertTrue( ConfigurationLoader::is_registered_block( 'remote-data-blocks/some_slick_block' ) );
 	}
 }
